fixing work label printing

+++ jacq-service/src/main/java/org/jacq/service/rest/impl/LabelServiceImpl.java
@@ -33,7 +33,7 @@ public class LabelServiceImpl implements LabelService, Serializable {
      * @see LabelService#getWork()
      */
     @Override
-    public Response getWork(int botanicalObjectId) {
+    public Response getWork(long botanicalObjectId) {
         try {
             return labelManager.getWork(botanicalObjectId);
         } catch (Exception e) {
